startup and teardown should go into the main loop
