+++ cheevos/cheevos.c
@@ -49,7 +49,6 @@
 #include "../tasks/tasks_internal.h"
 
 #include "../verbosity.h"
-#include "../debug/main.h"
 #include <rthreads/rthreads.h>
 
 /* Define this macro to prevent cheevos from being deactivated. */
@@ -364,26 +363,6 @@ static cheevos_locals_t cheevos_locals =
 bool cheevos_loaded      = false;
 int  cheats_are_enabled  = 0;
 int  cheats_were_enabled = 0;
-
-#ifdef HAVE_DEBUGGER
-typedef struct hunter
-{
-   void *empty;
-   volatile bool quit;
-   slock_t *lock;
-   sthread_t *thread;
-   bool deinit;
-} hunter_t;
-
-static void hunter_thread(void *data)
-{
-   hunter_t *handle = (hunter_t*)data;
-   debugger_init();
-   debugger_draw(&(handle->deinit));
-}
-
-hunter_t *handle;
-#endif
 
 /*****************************************************************************
 Supporting functions.
@@ -2643,11 +2622,6 @@ bool cheevos_unload(void)
 
    cheevos_loaded = 0;
 
-#ifdef HAVE_DEBUGGER
-   slock_lock(handle->lock);
-   handle->deinit = true;
-   slock_unlock(handle->lock);
-#endif
    return true;
 }
 
@@ -3668,17 +3642,5 @@ bool cheevos_load(const void *data)
 
    task_queue_push(task);
 
-#ifdef HAVE_DEBUGGER
-   handle = (hunter_t*)calloc(1, sizeof(*handle));
-
-   handle->lock         = slock_new();
-   handle->thread       = sthread_create(hunter_thread, handle);
-   if (!handle->thread)
-   {
-      slock_free(handle->lock);
-      free(handle);
-      return NULL;
-   }
-#endif
    return true;
 }
